requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
celery>=5.3.4
redis>=5.0.0
SQLAlchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.7
pydantic>=2.3.0
pydantic-settings>=2.0.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.0
beautifulsoup4>=4.12.2
pytest>=7.4.2
httpx>=0.25.0
aiohttp>=3.9.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# LangGraph
langgraph>=0.0.10
langchain-core>=0.1.0

# Medium publishing (Playwright)
playwright>=1.40.0

# Google Trends
pytrends>=4.9.0

# Google Gemini (image generation fallback)
google-genai>=1.0.0
